Add exercise 03 and update REST tasks

+++ UebungsAngabe02/4xHIF-Plfs03-POS/src/Aufgabe3_RestfulApi/Dtos/EmployeeWorkloadDto.cs
@@ -7,7 +7,7 @@ public record EmployeeWorkloadDto(
     string OfficeCountry,
     int OpenTaskCount,
     int ActiveProjectCount,
-    int BookedHours,
-    int AvgHoursPerTask,
+    decimal BookedHours,
+    decimal AvgHoursPerTask,
     List<string> Skills
     );
